Consolidate API types
