--- src/@types/api.ts
@@ -1,4 +1,0 @@
-export interface TitleStatsResponse {
-  playerCount: number;
-  battleCount: number;
-}
