--- src/@types/daily-api.ts
@@ -1,10 +1,0 @@
-import type { ScoreboardCategory } from "#enums/scoreboard-category";
-
-export interface GetDailyRankingsRequest {
-  category: ScoreboardCategory;
-  page?: number;
-}
-
-export interface GetDailyRankingsPageCountRequest {
-  category: ScoreboardCategory;
-}
