--- src/@types/ranking-entry.ts
@@ -1,6 +1,0 @@
-export interface RankingEntry {
-  rank: number;
-  username: string;
-  score: number;
-  wave: number;
-}
