--- src/@types/system-savedata-api.ts
@@ -1,20 +1,0 @@
-import type { SystemSaveData } from "#types/system-data";
-
-export interface GetSystemSavedataRequest {
-  clientSessionId: string;
-}
-
-export class UpdateSystemSavedataRequest {
-  clientSessionId: string;
-  trainerId?: number;
-  secretId?: number;
-}
-
-export interface VerifySystemSavedataRequest {
-  clientSessionId: string;
-}
-
-export interface VerifySystemSavedataResponse {
-  valid: boolean;
-  systemData: SystemSaveData;
-}
