--- src/@types/user-info.ts
@@ -1,7 +1,0 @@
-export interface UserInfo {
-  username: string;
-  lastSessionSlot: number;
-  discordId: string;
-  googleId: string;
-  hasAdminRole: boolean;
-}
