+++ src/plugins/api/daily-api.ts
@@ -1,6 +1,5 @@
 import { ApiBase } from "#api/api-base";
-import type { GetDailyRankingsPageCountRequest, GetDailyRankingsRequest } from "#types/daily-api";
-import type { RankingEntry } from "#types/ranking-entry";
+import type { GetDailyRankingsPageCountRequest, GetDailyRankingsRequest, RankingEntry } from "#types/api-types";
 
 /**
  * A wrapper for daily-run API requests.
