+++ src/plugins/api/savedata-api.ts
@@ -2,7 +2,7 @@ import { ApiBase } from "#api/api-base";
 import { SessionSavedataApi } from "#api/session-savedata-api";
 import { SystemSavedataApi } from "#api/system-savedata-api";
 import { MAX_INT_ATTR_VALUE } from "#constants/game-constants";
-import type { UpdateAllSavedataRequest } from "#types/savedata-api";
+import type { UpdateAllSavedataRequest } from "#types/api-types";
 
 /**
  * A wrapper for savedata API requests.
