+++ test/plugins/api/account-api.test.ts
@@ -2,7 +2,7 @@ import { AccountApi } from "#api/account-api";
 import { SESSION_ID_COOKIE } from "#constants/app-constants";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { AccountInfoResponse } from "#types/account-api";
+import type { AccountInfoResponse } from "#types/api-types";
 import * as AppUtils from "#utils/app-utils";
 import { removeCookie, setCookie } from "#utils/app-utils";
 import { HttpResponse, http } from "msw";
