+++ test/plugins/api/savedata-api.test.ts
@@ -1,7 +1,7 @@
 import { SavedataApi } from "#api/savedata-api";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { UpdateAllSavedataRequest } from "#types/savedata-api";
+import type { UpdateAllSavedataRequest } from "#types/api-types";
 import { HttpResponse, http } from "msw";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
