+++ test/plugins/api/session-savedata-api.test.ts
@@ -1,7 +1,6 @@
 import { SessionSavedataApi } from "#api/session-savedata-api";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { SessionSaveData } from "#types/session-data";
 import type {
   ClearSessionSavedataRequest,
   ClearSessionSavedataResponse,
@@ -9,7 +8,8 @@ import type {
   GetSessionSavedataRequest,
   NewClearSessionSavedataRequest,
   UpdateSessionSavedataRequest,
-} from "#types/session-savedata-api";
+} from "#types/api-types";
+import type { SessionSaveData } from "#types/session-data";
 import { HttpResponse, http } from "msw";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
