+++ test/plugins/api/system-savedata-api.test.ts
@@ -1,13 +1,13 @@
 import { SystemSavedataApi } from "#api/system-savedata-api";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { SystemSaveData } from "#types/system-data";
 import type {
   GetSystemSavedataRequest,
   UpdateSystemSavedataRequest,
   VerifySystemSavedataRequest,
   VerifySystemSavedataResponse,
-} from "#types/system-savedata-api";
+} from "#types/api-types";
+import type { SystemSaveData } from "#types/system-data";
 import { HttpResponse, http } from "msw";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
